+++ src/package.json
@@ -11,7 +11,8 @@
     "core-js": "^3.6.5",
     "d3": "^6.2.0",
     "element-ui": "^2.13.2",
-    "vue": "^2.6.11"
+    "vue": "^2.6.11",
+    "vue-router": "^3.4.7"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
